pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.16",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "langchain-openai>=0.3.30",
    "langchainhub>=0.1.21",
    "tiktoken>=0.11.0",
    "pydantic>=2.0.0",
    "pytest>=7.0.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.7.1",
    "langchain-chroma>=0.2.5",
    "pinecone-client>=3.0.0",
    "langchain-pinecone>=0.1.0",
    "weaviate-client>=4.0.0,<5.0.0",
    "langchain-weaviate>=0.0.1",
    "psutil>=5.9.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "click>=8.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
]
[tool.pytest.ini_options]
addopts = "-v --tb=short"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
# Ignore model classes from being collected as test classes
filterwarnings = [
    "ignore::pytest.PytestCollectionWarning"
]